Modify update form
