+++ src/app/dashboard/dashboard.module.ts
@@ -51,13 +51,13 @@ import { ListForumComponent } from './forum/list-forum/list-forum.component';
 import { ListCardComponent } from './forum/list-forum/list-card/list-card.component';
 import { CommentComponent } from './forum/list-forum/list-card/comment/comment.component';
 import { UserProfileCardComponent } from './forum/user-profile-card/user-profile-card.component';
+import { ReplyComponent } from './forum/list-forum/list-card/comment/reply/reply.component';
+import { CreateFormComponent } from './forum/create-form/create-form.component';
 // demand ad
 import { DemandAdComponent } from './demand-ad/demand-ad.component';
 import { CreateDemandAdComponent } from './demand-ad/create-demand-ad/create-demand-ad.component';
 import { ListDemandAdComponent } from './demand-ad/list-demand-ad/list-demand-ad.component';
-import { ReplyComponent } from './forum/list-forum/list-card/comment/reply/reply.component';
-import { FileUploadDirective } from './forum/file-upload.directive';
-import { UpdateFormComponent } from './forum/update-form/update-form.component';
+
 
 @NgModule({
   declarations: [
@@ -77,7 +77,7 @@ import { UpdateFormComponent } from './forum/update-form/update-form.component';
     ListDemandAdComponent,
     ReplyComponent,
     UserProfileCardComponent,
-    UpdateFormComponent,
+    CreateFormComponent,
   ],
   imports: [
     CommonModule,
