+++ src/app/dashboard/forum/create-form/create-form.component.html
@@ -1,9 +1,9 @@
-<nb-card>
+<nb-card item>
     <nb-card-body>
-        <div class="list-container">
+        <div class="list-container" >
             <div class="form-container">
 
-                <form [formGroup]="updateForm" novalidate (ngSubmit)="onCreate()">
+                <form [formGroup]="discussionForm" novalidate (ngSubmit)="onCreate()">
                     <label for="title" id="form-title"> Title </label>
                     <input type="text" class="form-control" formControlName="title" nbInput fullWidth fieldSize="small"
                         status="primary" placeholder="Title" />
@@ -16,10 +16,10 @@
                         <nb-checkbox status="info" (change)="showFarmer = !showFarmer" [checked]="showFarmer">
                             For farmers
                         </nb-checkbox>
-                        <nb-checkbox status="info" (change)="showBuyer = !showBuyer" [checked]="showBuyer" style="margin-left: 10px;">
+                        <nb-checkbox status="info" (change)="showBuyer = !showBuyer" [checked]="showBuyer"
+                            style="margin-left: 10px;">
                             For buyers
                         </nb-checkbox>
-
                         <ngx-dropzone (change)="onSelect($event)" class="dropzone" [expandable]="true">
                             <ngx-dropzone-label>drop files </ngx-dropzone-label>
                             <ngx-dropzone-image-preview ngProjectAs="ngx-dropzone-preview" *ngFor="let image of images"
@@ -29,12 +29,12 @@
                         </ngx-dropzone>
 
                         <button nbButton shape="round" size="small" status="success" class="float-right"
-                            [disabled]="!updateForm.valid" style="margin-top: 10px;">
+                            [disabled]="!discussionForm.valid" style="margin-top: 10px;">
                             Submit
                         </button>
-
                     </div>
                 </form>
+
             </div>
         </div>
     </nb-card-body>
