+++ src/app/dashboard/forum/forum.component.html
@@ -5,54 +5,21 @@
   </div>
 
   <div class="list-container">
-    <div class="form-container">
-      <nb-accordion>
-        <nb-accordion-item #item>
-          <nb-accordion-item-body>
-            <form [formGroup]="discussionForm" novalidate (ngSubmit)="onCreate()" >
-              <label for="title" id="form-title"> Title </label>
-              <input type="text" class="form-control" formControlName="title" nbInput fullWidth fieldSize="small"
-                status="primary" placeholder="Title" required />
-              <label for="description" id="form-title" style="margin-top: 20px;">
-                Description
-              </label>
-              <textarea formControlName="des" nbInput fullWidth status="primary" placeholder="Description"></textarea>
-              <div class="form-bottom-container">
-                <nb-checkbox status="info" (change)="showFarmer = !showFarmer" [checked]="showFarmer">
-                  For farmers
-                </nb-checkbox>
-                <nb-checkbox status="info" (change)="showBuyer = !showBuyer" [checked]="showBuyer" style="margin-left: 10px;">
-                  For buyers
-                </nb-checkbox>
-
-                <ngx-dropzone (change)="onSelect($event)" class="dropzone" [expandable]="true">
-                  <ngx-dropzone-label>drop files </ngx-dropzone-label>
-                  <ngx-dropzone-image-preview ngProjectAs="ngx-dropzone-preview" *ngFor="let image of images" [file]="image" [removable]="true" (removed)="onRemove(image)">
-                    <ngx-dropzone-label>{{ image.name }} ({{ image.type }})</ngx-dropzone-label>
-                  </ngx-dropzone-image-preview>
-                </ngx-dropzone>
-
-                <button nbButton shape="round" size="small" status="success" class="float-right "
-                  style="margin-top: 10px;" [disabled]="!discussionForm.valid">
-                  Submit
-                </button>
-
-              </div>
-            </form>
-          </nb-accordion-item-body>
-        </nb-accordion-item>
-      </nb-accordion>
-      <hr>
+    <div class="form-container" #item>
+      <app-update-form *ngIf="isCreate" [postId]="postId" (hideForm)="toggleForm($event)"
+        [createOrUpdate]="createOrUpdate"></app-update-form>
     </div>
-    <div class="post-contaner" style="margin-top: -2%;">
+
+    <div class="post-contaner">
       <app-list-forum [showMyPost]="showMyPost"></app-list-forum>
     </div>
   </div>
 
   <div class="side-bar-continer">
-    <button nbButton status="primary" (click)="toggle()" id="add-btn">
-      Create New
+    <button nbButton status="primary" (click)="createPost()" id="add-btn">
+      {{ !isCreate ? 'Create New' : 'Cancel' }}
     </button>
+
     <nb-card size="medium" class="side-bar">
     </nb-card>
   </div>
