+++ src/app/dashboard/forum/forum.component.scss
@@ -13,6 +13,7 @@
 }
 .list-container{
   width: 800px;
+  margin-top: 15px;
   margin-left: 75px;
   display: flex;
   flex-direction: column;
@@ -27,18 +28,6 @@
   width: 120px;
   height: 40px;
   margin-left: 55px;
-}
-.validation-feedback {
-  color: red;
-}
-.form-bottom-container{
-  margin-left: 10px;
-  margin-top: 10px;
-}
-#form-title {
-  margin-left: 5px;
-  font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande',
-    'Lucida Sans', Arial, sans-serif;
 }
 .side-bar-continer{
   width: 500px;
